Got the initial state cleaned up for starting the game. Added func disableAllButtonsForBlueComputerPlayerStart (myGridArray: Game) which is only run when the blue player is set to computer. it does two things:
* sets the title of every button to ""
* disables every button

if blue is changed back to human, then enableOtherButtonsDuringMove(myGridArray: theGame) is run and all the buttons are re-enabled. we also run disableAllButtonsForBlueComputerPlayerStart() for a new game (if blue is set to computer) and for resizing, which is also a form of new game. since you can only set the player types before the first move is made, this works rather nicely.